[FE] #1 Add customer caching and retrieval functionality
Introduced caching functionality in the ISimtekService (Infrastructure) that caches customers details. In addition, introduced functions in the ICustomerRepository (Data Repository Abstractions) to retrieve customer data. Also, added the ability to transform DTO to Domain models in the CustomerMapperExtensions. Furthermore, a GetCustomersQueryHandler is introduced to handle the retrieval request. The file GetInterventionsQueryHandler was moved to a Handlers directory for better organization. The additional whitespace in SimtekDomain.csproj, SimtekMaui.Data.csproj and SimtekMaui.Application.csproj files is for better visibility. These additions and modifications contribute to the performance and code organization of the application.