#5 Add BottomSheet functionality in Add Customer
The commit adds functionality to show a BottomSheet utility in AddCustomerPage. To achieve this, `The49.Maui.BottomSheet` dependency is added in `SimtekMaui.csproj` and is utilized in `MauiProgram.cs`. A new `BottomInfo` component is created containing the layout and logic for the content to be displayed inside the BottomSheet. This component is then invoked in `AddCustomerPage.xaml.cs` when 'ShowBottomSheetEvent' is called. The trigger for this event is inside the `AddCustomerViewModel` which is attached to a Button with icon, placed inside `AddCustomerPage.xaml`.